compile with stricter standard compatibility. initial interface for
* libbalsa/{folder-scanners,misc,pop3,send}.c:
* libbalsa/mailbox_{maildir,mh}.c: compile with stricter standard compatibility.
* libbalsa/mailbox.[hc]: initial interface for server-side sorting.
* libbalsa/mailbox_imap.c: same.
* libbalsa/imap/imap-handle.c: improve handling of embedded msgs.
* src/main-window.c: use ssize_t when apriopriate.

+++ src/main-window.c
@@ -2015,7 +2015,7 @@ mail_progress_notify_cb()
     MailThreadMessage *threadmessage;
     MailThreadMessage **currentpos;
     void *msgbuffer;
-    uint count;
+    ssize_t count;
     gfloat fraction;
 
     msgbuffer = g_malloc(MSG_BUFFER_SIZE);
@@ -2144,7 +2144,7 @@ send_progress_notify_cb()
     SendThreadMessage *threadmessage;
     SendThreadMessage **currentpos;
     void *msgbuffer;
-    uint count;
+    ssize_t count;
     float fraction;
 
     msgbuffer = malloc(2049);
